% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabetizer.R
\name{alphabetizer}
\alias{alphabetizer}
\title{Alphabetizer Alphabetize using Unificado or Azümchefe order}
\usage{
alphabetizer(string, input = "AMU")
}
\arguments{
\item{string}{A character vector in Unificado or Azümchefe.}

\item{input}{A writing system such as "AMU" (Unificado) or "AZU" (Azümchefe)

The argument "AMU" corresponds to the alphabetical order followed
in \emph{Augusta, F. J. de(2017), Diccionario Mapudungún - Español. UCT
Ediciones}, therefore apostrophes must be used in the input text for
interdentals: t’, n’, l’

Defaults to AMU}
}
\description{
This functions allows you to sort a list of words by changing the letters
into a numeric sequence. For use, follow examples.
}
\examples{


list_amu <- c("ñaña", "n’ome", "mi", "ngürü", "nu")
list_azu <- c("ñaña", "nhome", "mi", "gürü", "nu")
data <- data.frame(list_azu, list_amu)
data[alphabetizer(list_amu),]
data[alphabetizer(list_azu, input = "AZU"),]

}
\author{
Aldo Berrios, \email{aldoberrios@gmail.cl}
}
