% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphemizer.R
\name{graphemizer}
\alias{graphemizer}
\title{Convert any text in Mapudungun between Mapuche alphabets}
\usage{
graphemizer(string, input = "AMU", output = "AZU")
}
\arguments{
\item{input}{Script in which the original text is written. Defaults to AMU

AMU = Alfabeto Mapuche Unificado
RAG = Alfabeto Raguileo
AZU = Alfabeto Azümchefe}

\item{output}{Script in which the output text should be written. Defaults to AZU.}
}
\description{
This functions changes a text written in any Mapuche alphabet into another
alphabet. Mapuche alphabets supported: Alfabeto Mapuche Unificado,
Raguileo and Azümchefe.
}
\examples{

graphemizer("Tüfachi domo l’awen’tukey tañi küme tremoam.", output = "AZU")
> [1] "Tüfachi zomo lhawenhtukey tañi küme txemoam."

graphemizer("Kiñe mapuce wenxu nvniey kiñe caja tañi kuwv mew.", input = "RAG", output = "AZU")
> [1] "Kiñe mapuche wentxu nüniey kiñe challa tañi kuwü mew."

# Several dental graphemes are supported in AMU:
graphemizer(c("pal’u", "pal'u", "paḻu", "pal̯u"), output = "RAG")
> [1] "pabu" "pabu" "pabu" "pabu"

graphemizer(c("pal’u", "pal'u", "paḻu", "pal̯u"), output = "AZU")
[1] "palhu" "palhu" "palhu" "palhu"

}
\author{
Aldo Berrios, \email{aldoberrios@gmail.cl}
}
