% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabizer.R
\name{syllabizer}
\alias{syllabizer}
\title{syllabizer separates an IPA text into syllables}
\usage{
syllabizer(string)
}
\arguments{
\item{string}{A character vector.}
}
\description{
'syllabizer' - A function that separates Mapudungun words written in the
International Phonetic Alphabet into syllables. It makes use of the
straigthforward (C)V(C) syllabic pattern of Mapudungun to find
syllable boundaries. It does not parse texts written
in the Alfabeto Mapuche Unificado, so the function phonemizer() must be
used before, in order to obtain an IPA text.
}
\examples{
# Change any text into IPA with phonemizer():
text <- phonemizer("trafmapuche")

# Separate the syllables using syllabizer():
syllabizer(text)
> [1] "ʈ͡ʂaf.ma.pu.t͡ʃe"

}
\author{
Aldo Berrios, \email{aldoberrios@gmail.cl}
}
